Restore mvvm pattern
